The Windows port is now fairly complete. Keymaps might not be complete, as I don't know all functionality of the editor to test them.
It seems that the old code base is working under the assumption that terminals are of fixed size. There are compile time constants `NROW`,
and `NCOL` which necessarily don't match the runtime variables `nrow` and `ncol`. The Windows port will use the compile time constants,
and initialize the runtime variables with corresponding values, and then resize the terminal to match those values. This way screen updates
will be correct. The other option would be to change the `NROW` and `NCOL` compile time constants into runtime variables, but that would
alter the original source profoundly which is not desirable.

The shell spawn works much like the UNIX counterpart. The main difference between UNIX and Windows is that there is no "user selected shell".
Instead, there's a system shell, and that's it. I have added the possibility for the user to override this with for example powershell, and
that is done with the environment variable `UEMACS_SHELL`. If that variable does not exist, then uemacs will check if the system shell is set
(`ComSpec` environment variable). If that's not defined, then it'll shoot from the hip and execute `cmd.exe`.